requirements.txt
actionlib==1.14.0
angles==1.9.13
anyio==4.4.0
astor==0.8.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.14
beautifulsoup4==4.12.3
blinker==1.8.2
bondpy==1.8.6
cachetools==5.3.3
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cssselect==1.2.0
cssutils==2.11.1
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dill==0.3.8
dynamic-reconfigure==1.7.3
easyocr==1.7.1
enum34==1.1.10
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
fire==0.6.0
flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
fsspec==2024.5.0
future==1.0.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
image-geometry==1.16.2
imageio==2.34.1
imgaug==0.4.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
interactive-markers==1.12.0
itsdangerous==2.2.0
jinja2==3.1.4
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.5
laser-geometry==1.6.7
lazy-loader==0.4
Levenshtein==0.25.1
lmdb==1.4.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.16.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.4.0.46
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.82
openpyxl==3.1.3
opt-einsum==3.3.0
packaging==24.0
paddle-bfloat==0.1.7
paddleocr==2.3
paddlepaddle==2.4.2
pandas==2.0.3
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
protobuf==3.20.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyniryo==1.1.2
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-Levenshtein==0.25.1
python-qt-binding==0.4.4
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
rapidfuzz==3.9.3
rarfile==4.2
requests==2.31.0
resource-retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.25
scikit-image==0.17.2
scipy==1.10.1
seaborn==0.13.2
sensor-msgs==1.13.1
shapely==2.0.4
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
tifffile==2023.7.10
topic-tools==1.16.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing-extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
urllib3==2.2.1
visualdl==2.5.3
werkzeug==3.0.3
xacro==1.14.18
zipp==3.18.1
